+++ datacrunch_compat/datacrunch/datacrunch.py
@@ -38,7 +38,7 @@ __all__ = [
 import warnings
 
 warnings.warn(
-    'datacrunch import is deprecated; use verda instead.',
+    'datacrunch is deprecated; use verda package instead: https://github.com/verda-cloud/sdk-python/blob/master/CHANGELOG.md#1170---2025-11-26',
     DeprecationWarning,
     stacklevel=2,
 )